Refactored status management
